// src/products/domain/models/brand.entity.ts
/**
 * Brand interface
 *
 * @export
 * @interface Brand
 */
export interface Brand {
  /**
   * Brand id
   *
   * @type {number}
   * @memberof Brand
   */
  id: number;
  /**
   * Brand name
   *
   * @type {string}
   * @memberof Brand
   */
  name: string;
  /**
   * Brand image
   *
   * @type {string}
   * @memberof Brand
   */
  image: string;
}